requirements.txt
# ─── Backend Framework ─────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic[email]==2.9.0
email-validator>=2.0.0
python-multipart==0.0.9
python-dotenv==1.0.1

# ─── MongoDB (async driver) ───────
motor==3.6.0
pymongo==4.9.0

# ─── Security ──────────────────────
bcrypt>=4.2.0
PyJWT>=2.9.0
slowapi>=0.1.9
bleach>=6.1.0
Pillow>=10.0.0
reportlab>=4.0.0

# ─── AI / Computer Vision ─────────
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# ─── HTTP Client (AI → Backend) ───
aiohttp>=3.9.0
